rename gtidy3d to tidy3d
