+++ docs/notebooks/materials.py
@@ -11,7 +11,7 @@ import numpy as np
 from gdsfactory.components.taper import taper_sc_nc
 from gdsfactory.pdk import Pdk
 
-import gplugins.gtidy3d as gt
+import gplugins.tidy3d as gt
 
 gf.config.rich_output()
 PDK = gf.generic_tech.get_generic_pdk()
