+++ gplugins/photonic_circuit_models/mzi.py
@@ -39,7 +39,7 @@ def mzi(
 if __name__ == "__main__":
     import matplotlib.pyplot as plt
 
-    import gplugins.gtidy3d as gt
+    import gplugins.tidy3d as gt
 
     nm = 1e-3
     strip = gt.modes.Waveguide(
