+++ gplugins/tidy3d/write_sparameters.py
@@ -24,8 +24,8 @@ from gdsfactory.typings import (
     Tuple,
 )
 
-from gplugins.gtidy3d.get_results import _executor, get_results_batch
-from gplugins.gtidy3d.get_simulation import get_simulation, plot_simulation
+from gplugins.tidy3d.get_results import _executor, get_results_batch
+from gplugins.tidy3d.get_simulation import get_simulation, plot_simulation
 from gplugins.utils import port_symmetries
 from gplugins.utils.get_sparameters_path import (
     get_sparameters_path_tidy3d as get_sparameters_path,
